Refactor cloud backup manager modules
Move cloud-only, other-backups, remote-inventory, and sync

wallet-change logic into dedicated files in cloud_backup_manager.
